requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
click==8.1.8
Django==4.2.18
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
line-profiler==4.2.0
multidict==6.1.0
packaging==24.2
playwright==1.48.0
pluggy==1.5.0
propcache==0.2.0
pydantic==2.10.5
pydantic-core==2.27.2
pyee==12.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-django==4.9.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.41.3
tomli==2.2.1
typing-extensions==4.12.2
uvicorn==0.33.0
yarl==1.15.2
